Update PAGA plots
