+++ scanpy/tests/test_plotting.py
@@ -1284,79 +1284,6 @@ def test_scatter_rep(tmpdir):
             assert "Error" in comp, f"{s1.outpth}, {s2.outpth} aren't supposed to match"
 
 
-def test_paga(image_comparer):
-    # Sometimes things shift a pixel or so, resulting in diffs up to ~27
-    # The 1px-edges aren’t that good actually as they’re ignored at this tol …
-    save_and_compare_images = image_comparer(ROOT, FIGS, tol=30)
-
-    pbmc = sc.datasets.pbmc68k_reduced()
-    sc.tl.paga(pbmc, groups='bulk_labels')
-
-    common = dict(threshold=0.5, max_edge_width=1.0, random_state=0, show=False)
-
-    # delete bulk_labels_colors to test the creation of color list by paga
-    del pbmc.uns['bulk_labels_colors']
-    sc.pl.paga(pbmc, **common)
-    save_and_compare_images('master_paga')
-
-    sc.pl.paga(pbmc, color='CST3', **common)
-    save_and_compare_images('master_paga_continuous')
-
-    pbmc.obs['cool_feature'] = pbmc[:, 'CST3'].X.squeeze()
-    sc.pl.paga(pbmc, color='cool_feature', **common)
-    save_and_compare_images('master_paga_continuous_obs')
-
-    sc.pl.paga(pbmc, color=['CST3', 'GATA2'], **common)
-    save_and_compare_images('master_paga_continuous_multiple')
-
-    sc.pl.paga_compare(pbmc, legend_fontoutline=2, **common)
-    save_and_compare_images('master_paga_compare')
-
-    sc.pl.paga_compare(pbmc, color='CST3', legend_fontsize=5, **common)
-    save_and_compare_images('master_paga_compare_continuous')
-
-    sc.pl.paga_compare(pbmc, basis='X_pca', legend_fontweight='normal', **common)
-    save_and_compare_images('master_paga_compare_pca')
-
-    colors = {
-        c: {cm.Set1(_): 0.33 for _ in range(3)}
-        for c in pbmc.obs["bulk_labels"].cat.categories
-    }
-    colors["Dendritic"] = {cm.Set2(_): 0.25 for _ in range(4)}
-
-    sc.pl.paga(pbmc, color=colors, colorbar=False)
-    save_and_compare_images('master_paga_pie')
-
-
-def test_paga_path(image_comparer):
-    save_and_compare_images = image_comparer(ROOT, FIGS, tol=15)
-
-    pbmc = sc.datasets.pbmc68k_reduced()
-    sc.tl.paga(pbmc, groups='bulk_labels')
-
-    pbmc.uns['iroot'] = 0
-    sc.tl.dpt(pbmc)
-    sc.pl.paga_path(
-        pbmc,
-        nodes=['Dendritic'],
-        keys=['HES4', 'SRM', 'CSTB'],
-        show=False,
-    )
-    save_and_compare_images('master_paga_path')
-
-
-def test_paga_compare(image_comparer):
-    # Tests that https://github.com/theislab/scanpy/issues/1887 is fixed
-    save_and_compare_images = image_comparer(ROOT, FIGS, tol=15)
-
-    pbmc = sc.datasets.pbmc3k_processed()
-    sc.tl.paga(pbmc, groups="louvain")
-
-    sc.pl.paga_compare(pbmc, basis="umap", show=False)
-
-    save_and_compare_images('master_paga_compare_pbmc3k')
-
-
 def test_no_copy():
     # https://github.com/theislab/scanpy/issues/1000
     # Tests that plotting functions don't make a copy from a view unless they
